DllImport 関連のコードを分離
